pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "postcanvas"
version         = "0.1.0"
description     = "Generate pixel-perfect social media images from Python Pydantic models"
readme          = "README.md"
license         = { text = "MIT" }
requires-python = ">=3.10"
keywords        = ["social media", "instagram", "image generation", "pillow", "pydantic"]
dependencies    = [
    "Pillow>=10.0",
    "pydantic>=2.0",
    "requests>=2.28",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["postcanvas*"]
